fix: rename my hunter
